pyproject.toml
[project]
name = "lelamp-runtime"
version = "0.1.0"
description = "Python-based control system for LeLamp robotic lamp with voice AI, motor control, and vision capabilities"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "feetech-servo-sdk>=1.0.0,<2.0.0",
    "lerobot>=0.3.3",
    "livekit>=1.0.23",
    "livekit-agents>=1.5.0",
    "livekit-plugins-openai>=1.5.0",
    "livekit-plugins-silero>=1.5.0",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "numpy>=2.2.6",
    "pvporcupine>=3.0.5,<4.0.0",
    "pvrecorder>=1.2.7,<2.0.0",
    "pyarrow==20.0.0",
    "pyaudio>=0.2.14,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "rpi-ws281x>=5.0.0,<6.0.0; sys_platform == 'linux'",
    "sounddevice>=0.5.2,<1.0.0",
    "httpx>=0.28.1,<1.0.0",
    "certifi>=2024.0.0",
    "packaging>=24.0,<25.0.0",
]

[tool.uv.sources]
lerobot = { git = "https://github.com/huggingface/lerobot", tag = "v0.3.3" }

[project.optional-dependencies]
hardware = [
    "adafruit-circuitpython-neopixel>=6.3.0,<7.0.0",
    "rpi-ws281x>=5.0.0,<6.0.0",
]
vision = [
    "mediapipe>=0.10.0,<1.0.0; (sys_platform == 'darwin' and platform_machine == 'arm64') or (sys_platform == 'linux' and platform_machine == 'x86_64') or (sys_platform == 'win32' and platform_machine == 'AMD64')",
    "opencv-python>=4.9.0,<5.0.0; sys_platform != 'linux'",
    "opencv-python-headless>=4.9.0,<5.0.0; sys_platform == 'linux'",
]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "pytest-mock>=3.12.0,<4.0.0",
    "ruff>=0.3.0,<1.0.0",
]
api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.6",
    "zeroconf>=0.132.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["lelamp*"]
exclude = ["tests*", "scripts*", "web*"]
